requirements.txt
flask
transformers
torch
numpy
sentencepiece
requests
python-dotenv
cryptography
spacy
sumy
scikit-learn
pandas
accelerate